Pinned a few plugin
Refactored a few templates

+++ templates/perl.PKGBUILD
@@ -1,4 +1,4 @@
-# Maintainer: dracorp aka Piotr Rogoza <piotr.r.public@gmail.com>
+# Maintainer: dracorp aka Piotr Rogoza <piotr.r.public at gmail.com>
 
 pkgname=perl-<+$LBASENAME$+>
 pkgver=<+#VERSION#+>
@@ -20,38 +20,37 @@ install=
 noextract=()
 options=(!emptydirs)
 source=("http://search.cpan.org/CPAN/authors/id/$_author/$_perlmod-$pkgver.tar.gz")
+unset PERL5LIB PERL_MM_OPT PERL_MB_OPT PERL_LOCAL_LIB_ROOT
+export PERL_MM_USE_DEFAULT=1 PERL_AUTOINSTALL=--skipdeps MODULEBUILDRC=/dev/null
 
 build(){
   cd "$srcdir"/$_perlmod-$pkgver
-  unset PERL5LIB PERL_MM_OPT PERL_MB_OPT PERL_LOCAL_LIB_ROOT
-  export PERL_MM_USE_DEFAULT=1 PERL_AUTOINSTALL=--skipdeps MODULEBUILDRC=/dev/null
 
-  # If using Makefile.PL
-  /usr/bin/perl Makefile.PL
-  make
-
-  # If using Build.PL
-  /usr/bin/perl Build.PL
-  ./Build
+  if [ -f Makefile.PL ]; then
+    /usr/bin/perl Makefile.PL
+    make
+  else
+    /usr/bin/perl Build.PL
+    ./Build
+  fi
 }
 check(){
   cd "$srcdir"/$_perlmod-$pkgver
 
-  unset PERL5LIB PERL_MM_OPT PERL_MB_OPT PERL_LOCAL_LIB_ROOT
-  export PERL_MM_USE_DEFAULT=1
-  # If using Makefile.PL
-  make test
-  # If using Build.PL
-  ./Build test
+  if [ -f Makefile.PL ]; then
+    make test
+  else
+    ./Build test
+  fi
 }
 package(){
   cd "$srcdir"/$_perlmod-$pkgver
 
-  unset PERL5LIB PERL_MM_OPT PERL_MB_OPT PERL_LOCAL_LIB_ROOT
-  # If using Makefile.PL
-  make install INSTALLDIRS=vendor DESTDIR="$pkgdir"
-  # If using Build.PL
-  ./Build install installdirs=vendor destdir="$pkgdir"
+  if [ -f Makefile.PL ]; then
+    make install INSTALLDIRS=vendor DESTDIR="$pkgdir"
+  else
+    ./Build install installdirs=vendor destdir="$pkgdir"
+  fi
 
   # remove perllocal.pod and .packlist
   #find "$pkgdir" -name .packlist -o -name perllocal.pod -delete
